graphics_generator/pyproject.toml
[project]
name = "graphics-generator"
version = "0.1.0"
description = "PNG to SystemVerilog sprite ROM generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0",
    "pillow>=10.0",
]

[project.scripts]
png2sv = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
